+++ solutions/day5.go
@@ -14,13 +14,13 @@ type Day5Input struct{
 
 func (s *Day5Input) Prepare(input string) {
 	s.VM = intcode.VM{}
-	s.Memory = make(map[int]int)
+	s.Memory = make(map[int64]int64)
 
 	for k,v := range strings.Split(input, ",") {
 		pi, err := strconv.ParseInt(v, 10, 64)
 		util.PanicIfErr(err)
 
-		s.Memory[k] = int(pi)
+		s.Memory[int64(k)] = pi
 	}
 
 	s.IoMgr = &intcode.PreparedIO{}
@@ -29,10 +29,10 @@ func (s *Day5Input) Prepare(input string) {
 func (s *Day5Input) Part1() string {
 	io := s.IoMgr.(*intcode.PreparedIO)
 	// input the ID needed, reset stdout
-	io.Stdin = []int{1}
-	io.Stdout = []int{}
+	io.Stdin = []int64{1}
+	io.Stdout = []int64{}
 	// Blacklist all part 2 functions
-	s.BlacklistedOps = map[int]bool{
+	s.BlacklistedOps = map[int64]bool{
 		5:true,
 		6:true,
 		7:true,
@@ -47,8 +47,8 @@ func (s *Day5Input) Part1() string {
 func (s *Day5Input) Part2() string {
 	io := s.IoMgr.(*intcode.PreparedIO)
 	// input the ID needed, reset stdout
-	io.Stdin = []int{5}
-	io.Stdout = []int{}
+	io.Stdin = []int64{5}
+	io.Stdout = []int64{}
 
 	s.Autorun()
